patternscanned

+++ src/hooker.h
@@ -22,7 +22,7 @@ struct dlinfo_t
 
 namespace Hooker
 {
-	bool GetLibraryInformation(const char* library, uintptr_t* address, size_t* size);
+	bool GetLibraryInformation(const char* library, unsigned int* address, size_t* size);
 	void InitializeVMHooks();
 	bool HookRecvProp(const char* className, const char* propertyName, std::unique_ptr<RecvPropHook>& recvPropHook);
 	void FindIClientMode();
